fix: remove imports left redundant by the Chart/plot split
CI (stack build --pedantic, i.e. -Wall -Werror) failed on the split
branch with unused-import errors. My local --pedantic run had passed
because it was INCREMENTAL: the modules whose code had moved to plot/
were served from cache and never recompiled, so GHC never re-emitted
the warnings. A clean build is the only check that means anything
after code is removed from a module.

A clean -Wall pass reported 90 warnings across 27 files: 86 redundant
imports (in src/ modules that lost their layout functions, and in the
generated plot/ modules that inherited whole import lists) plus 4
dead top-level binds -- lowerBound/upperBound/lowerArmNumerator/
upperArmNumerator duplicated into Plot.Payoffs.RangeAccrualNote, whose
live copies are in the pure module.

Verified from a full `stack clean --full`:
  - stack build --pedantic: exit 0, zero warnings
  - stack test: passes
  - all 49 committed PNGs regenerate byte-identically
  - src/ transitive Chart taint: 0 of 58 modules
  - core library build-depends: base, mwc-random, vector
