+++ plot/Plot/TickPath.hs
@@ -6,31 +6,11 @@ module Plot.TickPath
   ( tickPathLayout
   ) where
 
-import Control.Monad.ST (runST)
 import Data.Colour
 import Data.Colour.Names
-import Data.Word (Word32)
 import qualified Data.Vector as V
-import qualified Data.Vector.Mutable as MV
-import qualified Data.Vector.Unboxed as VU
 import Graphics.Rendering.Chart.Easy
-import Pricing.PriceDeformation (uniswapMinTick)
-import SqrtGrid
-  ( Tick
-  , SqrtPriceX96(..)
-  , pattern Q96
-  , sqrtPrice
-  , tickFromSqrtPriceX96
-  )
-import System.Random.MWC (initialize)
-import System.Random.MWC.Distributions (standard)
-import Volatility.VolTermStructure
-  ( Step(..)
-  , VolTermStructure(..)
-  , unInstantaneousVol
-  )
 import TickPath
-
 
 tickPathLayout :: TickPath -> Layout Double Double
 tickPathLayout path = execEC $ do
