+++ src/Payoffs/ReplicaDelta.hs
@@ -16,13 +16,12 @@ module Payoffs.ReplicaDelta
   , principalDelta
   ) where
 
-import Greeks.Delta (PayoffDelta(..), PriceDeltaX96(..), deltaOfPayoff)
+import Greeks.Delta (PayoffDelta(..), PriceDeltaX96(..))
 import Liquidity.LiquidityChunk (LiquidityChunk, chunkAmount0, chunkLiquidity, chunkTickLower, chunkTickUpper)
 import Panoptic.LegChunk (legChunk)
 import Panoptic.MintPlan (MintPlan(..), fourLegNumLegs)
 import Panoptic.NId (panopticTokenType)
-import Payoffs.VolatilityReplica (fourLegReplica)
-import SqrtGrid (PayoffX96(..), SqrtPlot, SqrtPriceX96(..), mulDiv, pattern Q96, sqrtPriceX96)
+import SqrtGrid (PayoffX96(..), SqrtPriceX96(..), mulDiv, pattern Q96, sqrtPriceX96)
 
 -- | ∂_P of the chunk principal at p: token0 held = L·(b − p̄)·Q96/(p̄·b), p̄ = clamp(p; a, b).
 -- Staged as `getAmount0ForLiquidity`: mulDiv(L << 96, b − p̄, b) / p̄.
